Added configuration file to convert Automl EfficientDet model
